% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgpr_openclose.R
\name{RGP_close_conn}
\alias{RGP_close_conn}
\title{Closes a connection to RGP}
\usage{
RGP_close_conn(dbRGP)
}
\arguments{
\item{dbRGP}{handle returned by RGP_open_conn}
}
\description{
Closes a dbRGP handle
}
