% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RGP_get_customers.R
\name{RGP_get_customers}
\alias{RGP_get_customers}
\title{Get Customer Data}
\usage{
RGP_get_customers(
  user = Sys.getenv("RGP_USER"),
  password = Sys.getenv("RGP_PASS"),
  RGP_locations = NULL,
  return_all = FALSE
)
}
\arguments{
\item{user}{username of database on RGP server}

\item{RGP_locations}{output from a previous call to RGP_locations}

\item{return_all}{TRUE/FALSE whether to return all columns}

\item{pass}{password of 'user'}
}
\value{
all_customers a tibble
}
\description{
Gets primary data from customer data table using information on the
facilities obtained via a call to RGP_get_gyms. If multiple codes are
submitted, then this will loop over all locations and concatenate the results
}
