% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgpr_utilities.R
\name{RGP_get_gyms}
\alias{RGP_get_gyms}
\title{Retrieves list of all facilities}
\usage{
RGP_get_gyms(dbRGP)
}
\arguments{
\item{dbRGP}{connection handle from open_RGP_conn}
}
\value{
RGP_sites a tibble of information on the RGP sites
}
\description{
User supplies one dbRGP handle, and retrieves information for other RGP locations
}
