% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgpr_openclose.R
\name{RGP_Locations}
\alias{RGP_Locations}
\title{Retrieves list of all facilities}
\usage{
RGP_Locations(
  address = Sys.getenv("RGP_ADDRESS"),
  user = Sys.getenv("RGP_USER"),
  password = Sys.getenv("RGP_PASS"),
  db_name = Sys.getenv("RGP_DATABASE")
)
}
\arguments{
\item{dbRGP}{connection handle from open_RGP_conn}
}
\value{
RGP_sites is a list of information about the install
}
\description{
User supplies one dbRGP handle, and retrieves information for other RGP locations
}
