% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgpr_utilities.R
\name{RGP_open_conn}
\alias{RGP_open_conn}
\title{Opens a connection to RGP}
\usage{
RGP_open_conn(
  address = Sys.getenv("RGP_ADDRESS"),
  user = Sys.getenv("RGP_USER"),
  password = Sys.getenv("RGP_PASS"),
  db_name = Sys.getenv("RGP_DATABASE")
)
}
\arguments{
\item{address}{URL of the RGP server}

\item{user}{username of database on RGP server}

\item{db_name}{database name of RGP server}

\item{pass}{password of 'user'}
}
\value{
dbRGP
}
\description{
Opens connection to an RGP instance
}
