% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RGPr_openclose.R
\name{RGPr_close_conn}
\alias{RGPr_close_conn}
\title{Closes a connection to RGP}
\usage{
RGPr_close_conn(dbRGP)
}
\arguments{
\item{dbRGP}{handle returned by RGP_open_conn}
}
\description{
Closes a dbRGP handle
}
