% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RGPr_customer_guid.R
\name{RGPr_customer_guid}
\alias{RGPr_customer_guid}
\title{Get GUID-CUSTOMER_ID table}
\usage{
RGPr_customer_guid(RGPconn = NULL)
}
\arguments{
\item{RGPconn}{this is a connection to an RGP database from RGPr_open_conn}
}
\value{
a tibble containing GUID and CUSTOMER_ID for the current facility
}
\description{
This allows all calls, including calls for checkins and invoices, to return GUID
}
