% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RGPr_get_invoices.R
\name{RGPr_get_invoices}
\alias{RGPr_get_invoices}
\title{Get Invoices}
\usage{
RGPr_get_invoices(
  RGP_databases,
  start_date = NULL,
  end_date = NULL,
  return_all = FALSE
)
}
\arguments{
\item{RGP_databases}{output from a previous call to RGPr_Import_Gyms}

\item{start_date}{pull checkins after this date}

\item{end_date}{pull checkins before this date}

\item{return_all}{TRUE/FALSE whether to return all columns}
}
\value{
all_customers a tibble
}
\description{
Gets primary data from customer data table using information on the
facilities obtained via a call to RGP_get_gyms. If multiple codes are
submitted, then this will loop over all locations and concatenate the results
}
