% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RGPr_openclose.R
\name{RGPr_Import_Gyms}
\alias{RGPr_Import_Gyms}
\title{Retrieves list of all facilities}
\usage{
RGPr_Import_Gyms(filename)
}
\arguments{
\item{filename}{connection handle from open_RGP_conn}
}
\value{
RGP_sites is a list of information about the install
}
\description{
Imports information about each RGP install from a user-supplied csv file of the form
# First, you'll need to supply a file of gym database info in csv format
like this:
}
\details{
TAG , ADDRESS          , USER  ,    PASSWORD    , DBNAME ,      X     ,   Y
VAO , rgpdb.domain.com ,  bob  ,  bobs-password ,  va513 ,  -82.994614, 40.098792
VTC , 192.168.1.3      ,   bob ,  bobs-password ,  vatc  ,   -82.994928, 40.094341
}
