% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgpr_openclose.R
\name{RGPr_Import_Gyms}
\alias{RGPr_Import_Gyms}
\title{Retrieves list of all facilities}
\usage{
RGPr_Import_Gyms(filename)
}
\arguments{
\item{dbRGP}{connection handle from open_RGP_conn}
}
\value{
RGP_sites is a list of information about the install
}
\description{
User supplies one dbRGP handle, and retrieves information for other RGP locations
}
